Put entity-effects and -actions into their own folder
