+++ client-v2/src/app/store/entities/entities.actions.ts
@@ -1,10 +1,7 @@
 import { createActionGroup, emptyProps, props } from '@ngrx/store'
-import { HttpServerErrorResponse } from 'src/app/http/types'
 import { EntityPreview } from 'src/app/fullstack-shared-models/entities.model'
-import { CreateTasklistDto, TaskList } from 'src/app/fullstack-shared-models/list.model'
-import { CreateTaskDto, TaskPreview, TaskPriority, TaskStatus } from 'src/app/fullstack-shared-models/task.model'
+import { HttpServerErrorResponse } from 'src/app/http/types'
 import { EntityCrudDto } from 'src/app/services/entities.service'
-import { PartialRequired } from 'src/app/utils/type.helpers'
 
 export type HttpServerErrorResponseWithData<T = { id: string }> = HttpServerErrorResponse & T
 
@@ -33,53 +30,5 @@ export const entitiesActions = createActionGroup({
         delete: props<EntityCrudDto>(),
         'delete success': props<EntityCrudDto>(),
         'delete error': props<HttpServerErrorResponseWithData>(),
-    },
-})
-
-export const listActions = createActionGroup({
-    source: 'Entity/Lists',
-    events: {
-        'create task list': props<Partial<CreateTasklistDto>>(),
-        'create task list success': props<{ createdList: TaskList }>(),
-        'create task list error': props<HttpServerErrorResponse>(),
-
-        'update description': props<{ id: string; newDescription: string }>(),
-        'update description success': props<{ id: string; newDescription: string }>(),
-        'update description error': props<HttpServerErrorResponseWithData>(),
-
-        'duplicate list': props<{ id: string }>(),
-        'duplicate list success': props<{ id: string }>(),
-        'duplicate list error': props<{ id: string }>(),
-
-        'export list': props<{ id: string }>(),
-    },
-})
-
-export const taskActions = createActionGroup({
-    source: 'Entity/Tasks',
-    events: {
-        'load task previews': emptyProps(),
-        'load task previews success': props<{ previews: TaskPreview[] }>(),
-        'load task previews error': props<HttpServerErrorResponse>(),
-
-        'load root level tasks': props<{ listId: string }>(),
-        'load root level tasks success': props<{ listId: string; tasks: TaskPreview[] }>(),
-        'load root level tasks error': props<HttpServerErrorResponseWithData>(),
-
-        create: props<Partial<CreateTaskDto>>(),
-        'create success': props<{ createdTask: TaskPreview }>(),
-        'create error': props<HttpServerErrorResponseWithData>(),
-
-        'update status': props<{ id: string; status: TaskStatus }>(),
-        'update status success': props<{ id: string; status: TaskStatus }>(),
-        'update status error': props<HttpServerErrorResponseWithData>(),
-
-        'update priority': props<{ id: string; priority: TaskPriority }>(),
-        'update priority success': props<{ id: string; priority: TaskPriority }>(),
-        'update priority error': props<HttpServerErrorResponseWithData>(),
-
-        'update description': props<{ id: string; newDescription: string }>(),
-        'update description success': props<{ id: string; newDescription: string }>(),
-        'update description error': props<HttpServerErrorResponseWithData>(),
     },
 })
