+++ client-v2/src/app/store/entities/list/list.actions.ts
@@ -1,0 +1,23 @@
+import { createActionGroup, props } from '@ngrx/store'
+import { CreateTasklistDto, TaskList } from 'src/app/fullstack-shared-models/list.model'
+import { HttpServerErrorResponse } from 'src/app/http/types'
+import { HttpServerErrorResponseWithData } from '../entities.actions'
+
+export const listActions = createActionGroup({
+    source: 'Entity/Lists',
+    events: {
+        'create task list': props<Partial<CreateTasklistDto>>(),
+        'create task list success': props<{ createdList: TaskList }>(),
+        'create task list error': props<HttpServerErrorResponse>(),
+
+        'update description': props<{ id: string; newDescription: string }>(),
+        'update description success': props<{ id: string; newDescription: string }>(),
+        'update description error': props<HttpServerErrorResponseWithData>(),
+
+        'duplicate list': props<{ id: string }>(),
+        'duplicate list success': props<{ id: string }>(),
+        'duplicate list error': props<{ id: string }>(),
+
+        'export list': props<{ id: string }>(),
+    },
+})
